Fixed segfault: Pokémon with the same speed
When two Pokémon with the same speed used moves with the same priority,
a segfault would happen 25% of the time due to the randomness used in
the move prioritization.

To fix this, the move list is now shuffled and only then sorted
according to the priorities and speeds involved, without using any
random factors in the sorting phase.